# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yarn Dyeing Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/daily:
    get:
      tags: [reports]
      operationId: getDailyReport
      summary: Get daily order report
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Daily report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyReport'
  /reports/monthly:
    get:
      tags: [reports]
      operationId: getMonthlyReport
      summary: Get monthly order report
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
        - name: month
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyReport'
  /reports/factory:
    get:
      tags: [reports]
      operationId: getFactoryReport
      summary: Get factory-wise report
      parameters:
        - name: factoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Factory report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactoryReport'
...
  /users/{id}/role:
    patch:
      tags: [users]
      operationId: updateUserRole
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdate'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
components:
...
  schemas:
...
    Order:
      type: object
      required: [id, orderNo, buyerName, orderType, yarnType, color, quantityKg, status, receiveDate, createdAt]
      properties:
        id:
          type: integer
        orderNo:
          type: string
        buyerName:
          type: string
        factoryId:
          type: [integer, 'null']
        factoryName:
          type: [string, 'null']
        orderType:
          type: string
          enum: [Sample, Bulk]
        yarnType:
          type: string
        color:
          type: string
        quantityKg:
          type: number
        receiveDate:
          type: string
          format: date
        deliveryDate:
          type: [string, 'null']
          format: date
        status:
          type: string
          enum: [Received, Lab Dip, Dyeing Running, Finishing, Ready, Delivered]
        remarks:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    DailyReport:
      type: object
      required: [date, ordersReceived, totalKg, deliveredOrders]
      properties:
        date:
          type: string
          format: date
        ordersReceived:
          type: integer
        totalKg:
          type: number
        deliveredOrders:
          type: integer
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    MonthlyReport:
      type: object
      required: [year, month, sampleOrders, bulkOrders, totalKg, totalOrders]
      properties:
        year:
          type: integer
        month:
          type: integer
        sampleOrders:
          type: integer
        bulkOrders:
          type: integer
        totalKg:
          type: number
        totalOrders:
          type: integer
        deliveredOrders:
          type: integer
    FactoryReport:
      type: object
      required: [factoryId, factoryName, totalOrders, completedOrders, pendingOrders, totalKg]
      properties:
        factoryId:
          type: integer
        factoryName:
          type: string
        totalOrders:
          type: integer
        completedOrders:
          type: integer
        pendingOrders:
          type: integer
        totalKg:
          type: number
...
    AppUser:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: string
        name:
          type: [string, 'null']
        email:
          type: [string, 'null']
        role:
          type: string
          enum: [admin, manager, operator]
        profileImageUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, manager, operator]