# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yarn Dyeing Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface DailyReport {
  date: Date;
  ordersReceived: number;
  totalKg: number;
  deliveredOrders: number;
  orders?: Order[];
}